+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.2.5 - 7 Sep 2026
+
+Bug fix:
+- `decode` stop dropping optional fields of a container produced by a codec
+- `sanitize` now not applied to a codec decoded value
+
 # 1.2.3 - 6 Aug 2026
 
 Security fix: